requirements.txt
numpy==1.26.4
pandas==2.1.4
pytorch==2.1.1
pytorch-cuda==11.8
pytorch_lightning==2.1.3
datasets==2.16.1
spacy==3.7.2
nltk==3.9.1
tqdm==4.65.0
transformers==4.38.2
sentence-transformers==3.2.0
